All DAI-era commits squashed into one.
My first job was at DAI, where I first learned how to interact
with a shell and use dotfiles generally. This large commit is a
squash of all commits from that era that I squashed in December
2025, keeping only some of the commit messages from back then.
Originally, there were a lot of commits for small modifications,
file creations, etc...

------------------------------------------------------
Added Daily Notes (dn) shortcuts

dn writes a bullet-pointed string to a file with today's date in
YYYY-MM-DD format in the ~/dn/ folder.

dno does the same, but the first argument is the filename.
This can be used for future notes i.e. dno 2030-10-01 "I died".

dnt displays today's notes.

dnview displays all files, or when an argument like 2019-10 is passed, ~/dn/2019-10*.

https://github.com/tomlockwood/dn
------------------------------------------------------

- Made a bunch of changes to "dn". It now is "notas"
- Added time stamping to "notas"

------------------------------------------------------
Turns this repository into a utility to run on any machine

Cleaned up and updated to make this truly a utility repository I
can clone into any machine and run a config from.
-------------------------------------------------------