Added Daily Notes (dn) shortcuts
dn writes a bullet-pointed string to a file with today's date in YYYY-MM-DD format in the ~/dn/ folder.

dno does the same, but the first argument is the filename. This can be used for future notes i.e. dno 2030-10-01 "I died".

dnt displays today's notes.

dnview displays all files, or when an argument like 2019-10 is passed, ~/dn/2019-10*.

https://github.com/tomlockwood/dn