Adds a specific macOS setup file
I'm separating the large setup file into OS-specific files so that
I can tailor installations a bit more. Maybe once I'm done I merge
them again because they are so similar the separation doesn't make
that much sense, but for the construction right now this is best.

This commit is that first stage, with macOS. The setup file itself
is work in progress.